// src/types/database.ts
/**
 * This file was generated by kysely-codegen.
 * Please do not edit it manually.
 */

import type { ColumnType } from "kysely";

export type Generated<T> = T extends ColumnType<infer S, infer I, infer U>
  ? ColumnType<S, I | undefined, U>
  : ColumnType<T, T | undefined, T>;

export interface AndroidMetadata {
  locale: Generated<string | null>;
}

export interface WcedEntry {
  _id: Generated<number | null>;
  entry: string | null;
  head: string | null;
  page: number | null;
}

export interface WcedHead {
  _id: Generated<number | null>;
  entryid: number | null;
  head: string | null;
  metadata: Buffer | null;
  normalized_head: string | null;
  pos: string | null;
  type: string | null;
}

export interface WcedTranslation {
  _id: Generated<number>;
  entryid: number | null;
  translation: string | null;
}

export interface DB {
  android_metadata: AndroidMetadata;
  wced_entry: WcedEntry;
  wced_head: WcedHead;
  wced_translation: WcedTranslation;
}